# docs/swagger.yaml
definitions:
...
  handlers.CredentialGroupRequest:
    properties:
      credentials:
        items:
          $ref: '#/definitions/handlers.CredentialItem'
        type: array
      email:
        type: string
      group_name:
        type: string
    type: object
  handlers.CredentialItem:
    properties:
      cash_limit:
        type: number
      credential_id:
        type: string
    type: object
...
  models.DefaultResponseModel:
    properties:
      data:
        description: Use interface{} to allow any type for data
      message:
        type: string
      page:
        allOf:
        - $ref: '#/definitions/models.Pagination'
        description: The pagination information
      status_code:
        type: integer
      success:
        type: boolean
    type: object
  models.Pagination:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /oms/v1/credential-groups:
    get:
      description: Get all credential groups belonging to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DefaultResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DefaultResponseModel'
      security:
      - BearerAuth: []
      summary: Get credential groups by user ID
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a new credential group with multiple credentials (Admin
        only)
      parameters:
      - description: create credential group payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CredentialGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DefaultResponseModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DefaultResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DefaultResponseModel'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.DefaultResponseModel'
      security:
      - BearerAuth: []
      summary: Create credential group
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"